Safe conversion of strings to numbers, safe parsing dates as strings

+++ httpserver/ServerApplicationSettings.h
@@ -11,7 +11,7 @@ namespace HttpServer
 {
 	struct ServerApplicationSettings
 	{
-		int port;
+		long int port;
 		std::string root_dir;
 		std::string temp_dir;
 		size_t request_max_size;
